Added the second paddle and basic scorekeeping and ball resetting
The player one (left) paddle is now controlled with W and S, while the player two (right) paddle is now controlled with Up Arrow and Down Arrow. The ball now resets itself upon bounding out and updates the proper score on either side which are accompanied by large numbers in the background.